requirements.txt
numpy==1.26.4

ml-dtypes==0.2.0

tensorflow-macos==2.15.0
tensorflow-metal==1.1.0

opencv-python-headless==4.9.0.80

matplotlib==3.8.4
pillow==10.3.0
tqdm==4.66.4